+++ android/src/main/java/com/flurry/android/reactnative/FlurryModule.java
@@ -50,7 +50,7 @@ public class FlurryModule extends ReactContextBaseJavaModule {
     private static final String FLURRY_MESSAGING_EVENT = "FlurryMessagingEvent";
 
     private static final String ORIGIN_NAME = "react-native-flurry-sdk";
-    private static final String ORIGIN_VERSION = "3.0.0";
+    private static final String ORIGIN_VERSION = "3.1.0";
 
     private FlurryAgent.Builder mFlurryAgentBuilder;